Update issue 77
Request/Response services based on a SOAP --> JMS pattern can now be generated including sample transformers between XML and a CSV format.
Unit-tests classes are generated for the transformers and integration test is generated for the full service including negative tests for invalid input and timeouts.
Based on the sample-wsdl in soi-toolkits refapps.

git-svn-id: https://soi-toolkit.googlecode.com/svn/trunk@639 c6565b9c-4ce7-b367-afc1-6c09926acdf9