Update issue 92
PROBLEM WITH ERROR HANDLING, NULL IS RETURNED INSTEAD OF A SOAP-FAULT, SEE http://markmail.org/message/l33bwu4m4swtz3qf

ISSUE: Error handling, of for example timeouts, is not as expected until this issue is resolved, i.e. an empty response is returned instead of an expected soap-fault message.

git-svn-id: https://soi-toolkit.googlecode.com/svn/trunk@801 c6565b9c-4ce7-b367-afc1-6c09926acdf9